Break item header into separate file
