+++ packages/components/psammead-radio-schedule/src/ProgramCard/index.jsx
@@ -2,16 +2,8 @@ import React from 'react';
 import styled from '@emotion/styled';
 import { mediaIcons } from '@bbc/psammead-assets/svgs';
 import { GEL_SPACING, GEL_SPACING_DBL } from '@bbc/gel-foundations/spacings';
+import { C_METAL, C_WHITE } from '@bbc/psammead-styles/colours';
 import {
-  C_EBON,
-  C_KINGFISHER,
-  C_METAL,
-  C_POSTBOX,
-  C_SHADOW,
-  C_WHITE,
-} from '@bbc/psammead-styles/colours';
-import {
-  getSansBold,
   getSansRegular,
   getSerifMedium,
 } from '@bbc/psammead-styles/font-styles';
@@ -20,7 +12,6 @@ import {
   getMinion,
   getPica,
 } from '@bbc/gel-foundations/typography';
-import { Link } from '@bbc/psammead-story-promo';
 import {
   oneOf,
   oneOfType,
@@ -30,33 +21,9 @@ import {
   number,
 } from 'prop-types';
 import { scriptPropType } from '@bbc/gel-foundations/prop-types';
-import VisuallyHiddenText from '@bbc/psammead-visually-hidden-text';
-import LiveLabel from '@bbc/psammead-live-label';
-import {
-  formatUnixTimestamp,
-  formatDuration,
-} from '@bbc/psammead-timestamp-container/utilities';
-import detokenise from '@bbc/psammead-detokeniser';
-import durationDictionary from '../utilities';
-
-const TitleWrapper = styled.span`
-  color: ${({ titleColor }) => titleColor};
-  padding: ${GEL_SPACING} 0;
-  display: inline-block;
-  width: 100%;
-  ${({ service }) => service && getSansRegular(service)};
-  ${({ script }) => script && getPica(script)};
-`;
-
-const StyledLink = styled(Link)`
-  &:hover ${TitleWrapper} {
-    text-decoration: underline;
-  }
-
-  &:focus ${TitleWrapper} {
-    text-decoration: underline;
-  }
-`;
+import { formatDuration } from '@bbc/psammead-timestamp-container/utilities';
+import ScheduleItemHeader from '../ScheduleItemHeader';
+import { programStateConfig } from '../utilities';
 
 const CardWrapper = styled.div`
   padding-top: ${GEL_SPACING};
@@ -77,18 +44,6 @@ const StyledH3 = styled.h3`
   ${({ script }) => script && getPica(script)};
   color: ${({ headerTextColor }) => headerTextColor};
   margin: 0; /* Reset */
-`;
-
-const NextLabel = styled.span`
-  ${({ service }) => service && getSansBold(service)};
-  ${({ script }) => script && getPica(script)};
-  color: ${C_KINGFISHER};
-  display: inline-block;
-
-  ${({ dir }) =>
-    dir === 'rtl'
-      ? `margin-left: ${GEL_SPACING};`
-      : `margin-right: ${GEL_SPACING};`}
 `;
 
 const SummaryWrapper = styled.p`
@@ -129,120 +84,7 @@ const DurationWrapper = styled.time`
       : `padding-right: ${GEL_SPACING};`}
 `;
 
-const programStateConfig = {
-  live: {
-    backgroundColor: C_POSTBOX,
-    headerTextColor: C_EBON,
-    titleColor: C_SHADOW,
-    durationColor: C_WHITE,
-  },
-  next: {
-    backgroundColor: C_WHITE,
-    headerTextColor: C_METAL,
-    titleColor: C_METAL,
-    durationColor: C_KINGFISHER,
-  },
-  onDemand: {
-    backgroundColor: C_EBON,
-    headerTextColor: C_EBON,
-    titleColor: C_SHADOW,
-    durationColor: C_WHITE,
-  },
-};
-
-const renderHeaderContent = ({
-  props,
-  dir,
-  state,
-  link,
-  brandTitle,
-  startTime,
-  duration,
-}) => {
-  const {
-    nextLabel,
-    liveLabel,
-    listenLabelTranslations,
-    service,
-    script,
-    timezone,
-    locale,
-    linkComponent,
-    linkComponentAttr,
-    durationLabel,
-  } = props;
-
-  const isLive = state === 'live';
-  const isNext = state === 'next';
-
-  const formattedStartTime = formatUnixTimestamp({
-    timestamp: startTime,
-    format: 'HH:mm',
-    timezone,
-    locale,
-    isRelative: false,
-  });
-
-  const formattedDuration = detokenise(
-    durationLabel,
-    durationDictionary({ duration, locale }),
-  );
-
-  const episodeTitle = formatUnixTimestamp({
-    timestamp: startTime,
-    format: 'LL',
-    timezone,
-    locale,
-    isRelative: false,
-  });
-
-  const content = (
-    // eslint-disable-next-line jsx-a11y/aria-role
-    <span role="text">
-      <VisuallyHiddenText>{`${listenLabelTranslations[state]}, `}</VisuallyHiddenText>
-      {isLive && (
-        <LiveLabel
-          service={service}
-          dir={dir}
-          liveText={liveLabel}
-          ariaHidden
-        />
-      )}
-      {isNext && (
-        <NextLabel
-          aria-hidden="true"
-          service={service}
-          script={script}
-          dir={dir}
-        >
-          {`${nextLabel} `}
-        </NextLabel>
-      )}
-      {brandTitle}
-      <VisuallyHiddenText>, {formattedStartTime}, </VisuallyHiddenText>
-      <TitleWrapper
-        service={service}
-        script={script}
-        {...programStateConfig[state]}
-      >
-        {episodeTitle}
-      </TitleWrapper>
-      <VisuallyHiddenText>{`, ${formattedDuration} `}</VisuallyHiddenText>
-    </span>
-  );
-
-  const linkProps = { [linkComponentAttr]: link };
-
-  return state === 'next' ? (
-    content
-  ) : (
-    <StyledLink as={linkComponent} {...linkProps}>
-      {content}
-    </StyledLink>
-  );
-};
-
-const ProgramCard = ({ props, dir, program }) => {
+const ProgramCard = ({ dir, program, ...props }) => {
   const { service, script, locale } = props;
   const { state, startTime, link, brandTitle, summary, duration } = program;
   return (
@@ -253,15 +95,15 @@ const ProgramCard = ({ props, dir, program }) => {
           script={script}
           {...programStateConfig[state]}
         >
-          {renderHeaderContent({
-            props,
-            dir,
-            state,
-            link,
-            brandTitle,
-            startTime,
-            duration,
-          })}
+          <ScheduleItemHeader
+            {...props}
+            dir={dir}
+            state={state}
+            link={link}
+            brandTitle={brandTitle}
+            startTime={startTime}
+            duration={duration}
+          />
         </StyledH3>
         {summary && (
           <SummaryWrapper service={service} script={script}>
@@ -285,49 +127,34 @@ const ProgramCard = ({ props, dir, program }) => {
   );
 };
 
-const programCardPropTypes = {
+ProgramCard.propTypes = {
+  dir: oneOf(['rtl', 'ltr']),
+  program: {
+    state: string.isRequired,
+    duration: string.isRequired,
+    startTime: number.isRequired,
+    summary: string.isRequired,
+    link: string.isRequired,
+  }.isRequired,
+  durationLabel: string.isRequired,
   service: string.isRequired,
   script: shape(scriptPropType).isRequired,
   brandTitle: string.isRequired,
-  link: string.isRequired,
-  state: string.isRequired,
   nextLabel: string.isRequired,
   liveLabel: string.isRequired,
   listenLabelTranslations: string.isRequired,
-  startTime: number.isRequired,
   timezone: string,
   locale: string,
   linkComponent: oneOfType([elementType, string]),
   linkComponentAttr: string,
 };
 
-const programCardDefaultPropTypes = {
+ProgramCard.defaultProps = {
+  dir: 'ltr',
   timezone: 'Europe/London',
   locale: 'en-gb',
   linkComponent: 'a',
   linkComponentAttr: 'href',
-};
-
-renderHeaderContent.propTypes = {
-  ...programCardPropTypes,
-};
-
-renderHeaderContent.defaultProps = {
-  ...programCardDefaultPropTypes,
-};
-
-ProgramCard.propTypes = {
-  dir: oneOf(['rtl', 'ltr']),
-  duration: string.isRequired,
-  durationLabel: string.isRequired,
-  summary: string,
-  ...programCardPropTypes,
-};
-
-ProgramCard.defaultProps = {
-  dir: 'ltr',
-  summary: null,
-  ...programCardDefaultPropTypes,
 };
 
 export default ProgramCard;
